set directory
