+++ server.js
@@ -1,0 +1,6 @@
+const app = require('./src/app');
+
+const PORT = process.env.PORT || 3000;
+app.listen(PORT, () => {
+    console.log(`서버가 포트 ${PORT}에서 실행 중입니다`);
+});
