+++ src/routes/authRoutes.js
@@ -1,0 +1,9 @@
+const express = require('express');
+const { signup, login } = require('../controllers/authController');
+
+const router = express.Router();
+
+router.post('/signup', signup);
+router.post('/login', login);
+
+module.exports = router;
